+++ src/models/tipo_pedido.py
@@ -1,9 +1,9 @@
-from sqlalchemy import Column, Integer, String, Float, ForeignKey
+from sqlalchemy import Column, SmallInteger, String
 from src.models import Base, Session
 
 class TipoPedido(Base):
     __tablename__ = 'tipo_pedido'
-    id = Column(Integer, primary_key=True)
+    id = Column(SmallInteger, primary_key=True)
     descripcion = Column(String(30), unique=True, nullable=False )
     
     def __init__(self, descripcion):
